cloud-run/colab-worker/requirements.txt
flask==3.0.0
papermill==2.5.0
yfinance==0.2.36
pandas==2.2.0
matplotlib==3.8.3
scipy==1.12.0
numpy==1.26.4
google-cloud-storage==2.14.0
gunicorn==21.2.0
nbformat==5.9.2